requirements.txt
asgiref==3.4.1
Django==3.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
